Various data cleaning and visualization
Add a map illustrating the logsum value. Inspecting this figure encouraged me to change the distance calculation. We previously had tract centroid to park centroid; we now use population-weighted tract centroid to edge of park. For cases where the population-weighted centroid is *in* a park, we set the minimum distance to 0.1 miles to avoid infinite logsums.